requirements.txt
matplotlib>=3.0.2
Jinja2>=2.10
ujson>=1.35
numpy>=1.15.4
colour>=0.1.5
CherryPy>=18.1.0
Sphinx>=1.8.3
autodoc>=0.5.0
sphinx-markdown-builder>=0.4.1     
sphinx-rtd-theme>=0.4.3
pandas>=0.24.2
tmap-viz>=1.0.16 